Uses shadows
Big update, getting close to a version 1.1 where the chess engine is all done locally.

This update makes use of AWS Shadow states to store the FEN state of the chess game, and allows the chess rules to all be computed locally, making a much better user experience.

Known bugs:
- Broke checking for auto-update.