Code Cleanup + restart logic fix
Previously, when a game was finished, we'd redraw the difference between the end-game, and if all the pieces happened to be back on their default positions, try to restart a new game.

Now when a game finishes, the winning board will start a new game, and put the last game state into a new variable.  If the chessboard is in the previous game state, or the previous game 2nd last move state, it'll show that delta, otherwise, it'll get you to setup for a new game